# core-lending-mortgages-sdk/target/classes/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: core-lending-mortgages
  description: Mortgages Core Application
  contact:
    name: Firefly Software Solutions Inc
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.catalis.vc/mortgages
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
...
  /api/v1/mortgage-contracts/{mortgageContractId}/payment-schedule/{scheduleId}:
    get:
      tags:
        - MortgagePaymentSchedule
      summary: Get a specific payment schedule entry by ID
      operationId: getById_2
      parameters:
        - name: mortgageContractId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: scheduleId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MortgagePaymentScheduleDTO'
    put:
      tags:
        - MortgagePaymentSchedule
      summary: Update a payment schedule entry
      operationId: update_2
      parameters:
        - name: mortgageContractId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: scheduleId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MortgagePaymentScheduleDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MortgagePaymentScheduleDTO'
    delete:
      tags:
        - MortgagePaymentSchedule
      summary: Delete a payment schedule entry
      operationId: delete_2
      parameters:
        - name: mortgageContractId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: scheduleId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
...
  /api/v1/mortgage-contracts/{mortgageContractId}/payment-schedule:
    get:
      tags:
        - MortgagePaymentSchedule
      summary: List or search the mortgage payment schedule for a contract
      operationId: findAll_2
      parameters:
        - name: mortgageContractId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: pagination.pageNumber
          in: query
          description: Page number (0-based)
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: pagination.pageSize
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            format: int32
            default: 10
        - name: pagination.sortBy
          in: query
          description: Field to sort by
          required: false
          schema:
            type: string
        - name: pagination.sortDirection
          in: query
          description: Sort direction (ASC or DESC)
          required: false
          schema:
            type: string
            default: DESC
        - name: options.caseInsensitiveStrings
          in: query
          description: Enable case-insensitive string filtering
          required: false
          schema:
            type: boolean
            default: false
        - name: options.includeInheritedFields
          in: query
          description: Include fields from parent classes
          required: false
          schema:
            type: boolean
            default: false
        - name: filters.mortgageContractId
          in: query
          description: Exact filter for mortgage contract id
          required: false
          schema:
            type: string
        - name: filters.installmentNumber
          in: query
          description: Exact filter for installment number
          required: false
          schema:
            type: integer
            format: int32
        - name: rangeFilters.ranges[installmentNumber].from
          in: query
          description: Filter installment number from value
          required: false
          schema:
            type: integer
            format: int32
        - name: rangeFilters.ranges[installmentNumber].to
          in: query
          description: Filter installment number to value
          required: false
          schema:
            type: integer
            format: int32
        - name: filters.dueDate
          in: query
          description: Exact filter for due date
          required: false
          schema:
            type: string
        - name: filters.principalAmount
          in: query
          description: Exact filter for principal amount
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[principalAmount].from
          in: query
          description: Filter principal amount from value
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[principalAmount].to
          in: query
          description: Filter principal amount to value
          required: false
          schema:
            type: string
        - name: filters.interestAmount
          in: query
          description: Exact filter for interest amount
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[interestAmount].from
          in: query
          description: Filter interest amount from value
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[interestAmount].to
          in: query
          description: Filter interest amount to value
          required: false
          schema:
            type: string
        - name: filters.feeAmount
          in: query
          description: Exact filter for fee amount
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[feeAmount].from
          in: query
          description: Filter fee amount from value
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[feeAmount].to
          in: query
          description: Filter fee amount to value
          required: false
          schema:
            type: string
        - name: filters.escrowAmount
          in: query
          description: Exact filter for escrow amount
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[escrowAmount].from
          in: query
          description: Filter escrow amount from value
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[escrowAmount].to
          in: query
          description: Filter escrow amount to value
          required: false
          schema:
            type: string
        - name: filters.totalAmount
          in: query
          description: Exact filter for total amount
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[totalAmount].from
          in: query
          description: Filter total amount from value
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[totalAmount].to
          in: query
          description: Filter total amount to value
          required: false
          schema:
            type: string
        - name: filters.paymentStatus
          in: query
          description: Exact filter for payment status
          required: false
          schema:
            type: string
            enum:
              - SCHEDULED
              - PAID
              - OVERDUE
              - PARTIALLY_PAID
              - WAIVED
        - name: filters.paidDate
          in: query
          description: Exact filter for paid date
          required: false
          schema:
            type: string
        - name: filters.outstandingBalance
          in: query
          description: Exact filter for outstanding balance
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[outstandingBalance].from
          in: query
          description: Filter outstanding balance from value
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[outstandingBalance].to
          in: query
          description: Filter outstanding balance to value
          required: false
          schema:
            type: string
        - name: filters.lateFeeAmount
          in: query
          description: Exact filter for late fee amount
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[lateFeeAmount].from
          in: query
          description: Filter late fee amount from value
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[lateFeeAmount].to
          in: query
          description: Filter late fee amount to value
          required: false
          schema:
            type: string
        - name: filters.daysLate
          in: query
          description: Exact filter for days late
          required: false
          schema:
            type: integer
            format: int32
        - name: rangeFilters.ranges[daysLate].from
          in: query
          description: Filter days late from value
          required: false
          schema:
            type: integer
            format: int32
        - name: rangeFilters.ranges[daysLate].to
          in: query
          description: Filter days late to value
          required: false
          schema:
            type: integer
            format: int32
        - name: filters.paymentBreakdown
          in: query
          description: Exact filter for payment breakdown
          required: false
          schema:
            type: string
        - name: filters.createdAt
          in: query
          description: Exact filter for created at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[createdAt].from
          in: query
          description: Filter created at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[createdAt].to
          in: query
          description: Filter created at to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.updatedAt
          in: query
          description: Exact filter for updated at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[updatedAt].from
          in: query
          description: Filter updated at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[updatedAt].to
          in: query
          description: Filter updated at to value
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponseMortgagePaymentScheduleDTO'
    post:
      tags:
        - MortgagePaymentSchedule
      summary: Create a new payment schedule entry
      operationId: create_2
      parameters:
        - name: mortgageContractId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MortgagePaymentScheduleDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MortgagePaymentScheduleDTO'
...
components:
  schemas:
...
    MortgagePaymentScheduleDTO:
      required:
        - dueDate
        - installmentNumber
        - interestAmount
        - mortgageContractId
        - outstandingBalance
        - paymentStatus
        - principalAmount
        - totalAmount
      type: object
      properties:
        scheduleId:
          type: string
          format: uuid
          readOnly: true
        mortgageContractId:
          type: string
          format: uuid
        installmentNumber:
          maximum: 480
          minimum: 1
          type: integer
          format: int32
        dueDate:
          type: string
          format: date
        principalAmount:
          maximum: 999999.99
          exclusiveMaximum: false
          minimum: 0
          exclusiveMinimum: false
          type: number
        interestAmount:
          maximum: 999999.99
          exclusiveMaximum: false
          minimum: 0
          exclusiveMinimum: false
          type: number
        feeAmount:
          maximum: 99999.99
          exclusiveMaximum: false
          minimum: 0
          exclusiveMinimum: false
          type: number
        escrowAmount:
          maximum: 99999.99
          exclusiveMaximum: false
          minimum: 0
          exclusiveMinimum: false
          type: number
        totalAmount:
          maximum: 999999.99
          exclusiveMaximum: false
          minimum: 0.01
          exclusiveMinimum: false
          type: number
        paymentStatus:
          type: string
          enum:
            - SCHEDULED
            - PAID
            - OVERDUE
            - PARTIALLY_PAID
            - WAIVED
        paidDate:
          type: string
          format: date
        outstandingBalance:
          maximum: 999999999.99
          exclusiveMaximum: false
          minimum: 0
          exclusiveMinimum: false
          type: number
        lateFeeAmount:
          maximum: 99999.99
          exclusiveMaximum: false
          minimum: 0
          exclusiveMinimum: false
          type: number
        daysLate:
          maximum: 9999
          minimum: 0
          type: integer
          format: int32
        paymentBreakdown:
          maxLength: 2000
          minLength: 0
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    PaginationResponseMortgagePaymentScheduleDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: '#/components/schemas/MortgagePaymentScheduleDTO'
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
...